colisiones, resize, log print and movement

+++ source/RenderManager.h
@@ -2,6 +2,7 @@
 #include <SDL.h>
 #include <vector>
 #include "SceneManager.h"
+#include "InputManager.h"
 
 #define RM RenderManager::GetInstance()
 
@@ -13,12 +14,10 @@ private:
 	RenderManager(int width, int height, int flags);
 public:
 	RenderManager();
-	float windowWidth;
-	float windowHeight;
+	int windowWidth;
+	int windowHeight;
 	static RenderManager* GetInstance();
 	SDL_Renderer* GetRenderer();
 	void ClearScreen();
 	void RenderScreen();
-
-	
 };
